osutils: move FindGidOwning() to separate file
When building with CGO_ENABLED=0 group.go will not be picked up for the
build (the files are picked/rejected from as a whole). However, FindGidOwning()
does not require cgo. Moving it to separate file ensures that the call will be
available in non cgo builds as well.
